Update All_Performance_Tips.md
There is a lot of value in the material on and linked from this page. 

I have attempted to make it a more appealing read without over-explaining when there is a linked demo or doc page for that. I removed item 16 as it seemed to duplicate no 2.

Note I have referred to Konva as 'Konva' and not 'KonvaJS' - please feel free to correct this or anything else I got wrong!